requirements.txt
streamlit==1.31.0
pyairtable==2.2.1
pandas==2.1.4
plotly==5.18.0
python-dotenv==1.0.0
streamlit-plotly-events==0.0.6
mysql-connector-python==9.5.0
bcrypt==4.1.2
